% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_survey}
\alias{cat_survey}
\title{Cat Survey}
\format{
A data frame with 4316 rows and 151
variables:
\subsection{Demographic and background information}{

\describe{
\item{catid}{Numeric id of the cat.}
\item{breedgroup}{Breed group. See \code{breed_group} for more information.}
\item{sex}{Sex, takes values \code{"female"} and \code{"male"}.}
\item{age_behaviour}{Age when personality section was answered, in years.}
\item{other_cats}{Is there other cats living in the same household, takes
values \code{"Yes"} and \code{"no"}.}
\item{problematic_behavior}{Do you feel that your cat has
problematic/unwanted behaviors? Takes values \code{"no"}, \code{"a little"},
\code{"some"}, and \code{"a lot"}.}
}
}

\subsection{Personality and behavior}{

For all the statements below, the following likert scale was used:
\enumerate{
\item Strongly disagree
\item somewhat disagree
\item Neither agree or disagree
\item Somewhat agree
\item Strongly agree
}

NA values indicate "I don't know".

\describe{
\item{seeks_physical_contact_people}{Often seeks out physical contact from
people, nudges or nuzzles.}
\item{Opens_doors}{Opens doors and cabinets.}
\item{Rests_in_secure_places}{Rests or sleeps in secure places (e.g. in
laundry baskets or wardrobes).}
\item{Reacts_to_video_TV}{Reacts to the video or sound coming from the TV
or computer screen (e.g. nature documentaries, cat videos, mouse
cursor).}
\item{Refuses_litter_box_dirty}{Refuses to use the litter box if it is
dirty.}
\item{Freezes_examined_veterinarian}{Freezes when examined or treated by a
veterinarian.}
\item{Seems_confident}{Seems confident at home.}
\item{comfortable_social_gatherings}{Is comfortable and relaxed among
people in social gatherings (e.g. parties).}
\item{easily_scared_things_sounds}{Is easily scared even by small
unexpected things and sounds.}
\item{Defecates_inappropriate_places}{Defecates in inappropriate places
indoors.}
\item{Unexpectedly_scratches_bites}{Unexpectedly scratches or bites when
petted.}
\item{greets_unfamiliar_children}{Always greets unfamiliar children
visiting your home in a friendly manner (e.g. sniffs, rubs, purrs,
lies on the floor).}
\item{Enjoys_tearing_cardboard}{Enjoys tearing up cardboard and paper.}
\item{Runs_hides_noise_inside}{Runs or hides when hearing a sudden or
loud noise coming from inside the house (e.g. vacuum cleaner,
dropped object, instrument, doorbell).}
\item{Runs_hide_noise_outside}{Runs or hides when hearing a sudden or loud
noise coming from outside the house (e.g. jackhammer, fireworks).}
\item{Asks_in_out_rooms}{Asks to be let in and out of rooms (e.g. walks
toward a door and meows).}
\item{Freezes_nothing visible}{Freezes and stares intensely at nothing
visible.}
\item{Tries_wake_up}{Tries to wake up members of the household early in
the morning.}
\item{exhibits_bursts_running}{Often exhibits sudden bursts of running
(“zoomies”).}
\item{Talks_to_people}{Talks to people (e.g. meows when spoken to).}
\item{increased_activity_morning}{Shows increased activity in the early
hours of morning around dawn/sunrise.}
\item{destroys_left_alone}{Chews or destroys objects when left alone at
home.}
\item{Reacts_catnip}{Reacts strongly to catnip by, e.g. rubbing, rolling
over or licking.}
\item{Chases_animals}{Chases rodents, birds, reptiles, squirrels, rabbits,
insects or other small animals, given the opportunity.}
\item{purrs_petted}{Always purrs when petted.}
\item{scratch_bite_human_possession_prey}{Attempts to scratch or bite when
approached by human while in possession of prey, stolen or
especially delicious food.}
\item{comfortable_familiar_children}{Appears comfortable (confident,
relaxed) when playing with familiar children.}
\item{Growls_hisses_familiar_cat_staring}{Growls or hisses at a familiar
cat when the familiar cat is staring, growling, or hissing at
her/him.}
\item{scratch_bite_familiar_dogs}{Attempts to scratch or bite familiar
dogs.}
\item{Growls_hisses_examined_veterinarian}{Growls or hisses when examined
or treated by a veterinarian.}
\item{Growls_hisses_unfamiliar_dogs}{Growls or hisses at unfamiliar dogs.}
\item{lie_things_used_people}{Tends to lie on newspapers, books, keyboards
or other things that are being used by people.}
\item{excited_new_toys}{Gets excited about new toys.}
\item{Runs_while_playing}{Runs around in the house while playing.}
\item{Greets_unfamiliar_cats}{Greets unfamiliar adult cats in a friendly
manner (sniffs, touches noses).}
\item{Growls_hisses_familiar_dogs}{Growls or hisses at familiar dogs.}
\item{restlessness_home_modified}{Shows restlessness or fearfulness when
his/her home is modified (e.g. furniture moved from usual places,
fabrics or sheets changed).}
\item{eats_house_plants}{Chews or eats house plants or cut flowers when
given the opportunity (this does not include cat grass given for the
cat to eat).}
\item{eye_contact_familiar_people}{Makes eye contact with familiar people,
simultaneously squinting or slowly blinking its eyes.}
\item{Scratches_inappropriate_objects}{Scratches claws on inappropriate
objects or surfaces (e.g. couch, curtains, wallpaper) indoors.}
\item{scratch_bite_petted_base_tail}{Attempts to scratch or bite when
petted at the base of the tail.}
\item{hallway_household_leave}{Walks to the hallway/next to front door when
a household member prepares to leave the home.}
\item{not_want_use_same_litter_box}{Does not want to use the same litter
box than the other cats in the household. Choose "I don't know" if
he/she is the only cat in the household.}
\item{wrong_targeted_aggression}{If the target of the cat’s aggression is
inaccessible, the cat attempts to scratch or bite a “wrong”
accessible target (e.g. the cat sees an unfamiliar cat through the
window and bites a family member).}
\item{Enjoys_jumping_high}{Enjoys jumping and climbing to high places.}
\item{restlessness_member_household_affection_person}{Shows restlessness
(e.g. vocalizes) when a member of the household shows affection to a
person.}
\item{Gets_excite_small_animals_window}{Gets excited (e.g. makes chirping
or chattering noises, or lashes its tail) when looking at birds or
other small animals at the window.}
\item{scratch_bite_petted_belly}{Attempts to scratch or bite (in a
non-playful way) when petted on the belly.}
\item{Carries_toy}{Carries small objects or toys in his/her mouth.}
\item{listens_everything_you_say}{Attends and listens closely to
everything you say or do.}
\item{playing_chases_household_members}{While playing, always chases and
ambushes other household members (e.g. people, dogs, or cats)
playfully.}
\item{Growls_hisses_nails_clipped}{Growls or hisses when his/her nails are
clipped.}
\item{Growls_hisses_medicine_familiar}{Growls or hisses when given
medicine by a familiar person (this does not include vets).}
\item{Chases_imaginary_objects}{Chases or hunts imaginary objects.}
\item{scratch_bite_bathed}{Attempts to scratch or bite when being bathed.}
\item{Growls_hisses_bathed}{Growls or hisses when being bathed.}
\item{increased_activity_evening}{Shows increased activity in the evening
around dusk/sunset or late at night.}
\item{Runs_unfamiliar_cats}{Runs or hides from unfamiliar cats.}
\item{scratch_bite_brushed}{Attempts to scratch or bite when being
brushed.}
\item{Reacts_similarly_situations}{Reacts similarly in similar situations
(it is possible to predict how your cat will react).}
\item{Stares_unfamiliar_people}{Stares intensely (wide-eyed, not blinking)
at unfamiliar people.}
\item{self_mutilation}{Exhibits self-mutilation, e.g. pulls hairs off with
teeth, vigorously nibbles or bites his/her body parts.}
\item{investigates_new_objects}{Curiously investigates and explores new
objects or changes in his/her environment.}
\item{Growls_hisses_approached_familiar_cat_while_eating}{Growls or hisses
when approached by a familiar cat while eating.}
\item{scratch_bite_unfamiliar_dogs}{Attempts to scratch or bite unfamiliar
dogs.}
\item{restlessness_other_cats_fight}{Shows restlessness or irritation when
other cats in the family are having a fight.}
\item{excessive_grooming}{Shows excessive and intensive grooming
(inhibits other behaviors) throughout the day.}
\item{restlessness_unfamiliar_objects}{Shows restlessness or fearfulness
when unfamiliar objects are introduced into the home.}
\item{Plays_fetch}{Plays fetch with people (retrieves thrown objects or
toys).}
\item{Gets_along_with_cats}{Gets along well with other cats in the
household. Choose “I don't know” if he/she is the only cat in the
household.}
\item{Seeks_company_other_cats}{Seeks company of other cats in the
household. Choose "I don't know" if he/she is the only cat in the
household.}
\item{baby_cry_check}{When a baby is crying, cat goes to check the baby.}
\item{Growls_hisses_brushed}{Growls or hisses when being brushed.}
\item{Chase_tail}{Chases his/her own tail/rear end.}
\item{scratch_bite_examined_veterinarian}{Attempts to scratch or bite when
examined or treated by a veterinarian.}
\item{Rests_warm_appliances}{Rests or sleeps on top of warm appliances
(e.g.TV, printer, fireplace, radiator).}
\item{scratch_bite_nails_clipped}{Attempts to scratch or bite when his/her
nails are clipped.}
\item{greets_kittens}{Greets unfamiliar kittens in a friendly manner
(sniffs, touches noses).}
\item{Moves_elegantly}{Moves elegantly and gracefully.}
\item{scratch_bite_medicine_familiar_person}{Attempts to scratch or bite
when given medicine by a familiar person (this do not include vets).}
\item{Hides_household_member_leave}{Hides or escapes when a household
member prepares to leave the home.}
\item{Urinates_inappropriate}{Urinates (crouching position) in
inappropriate places.}
\item{Escapes_unfamiliar_people}{Escapes or hides from unfamiliar people.}
\item{Is_persevering}{Is persevering, persistent and determined. Does what
he/she wants even if he/she is told not to.}
\item{Stalks_moving_objects}{Stalks, chases or pounces on moving objects
(e.g. string, balls, soft toys).}
\item{meowing_left_alone}{Vocalizes by crying or meowing when left alone
(without people) at home.}
\item{Sprays_indoors}{Sprays (standing position with tail raised vertically
and vibrating) indoors (e.g. furniture, walls).}
\item{adapts_changes}{Readily adapts to changes in environment or daily
routines (e.g. changes in schedules, rearrangement of objects and
furniture).}
\item{Squirms_picked_held_arms}{Squirms and tries to escape when picked
up and held in arms.}
\item{Brings_prey_animals}{Brings prey animals into the home. If your cat
has not opportunity to do this, choose “I don't know.}
\item{Enjoys_playing_cats}{Enjoys playing with other cats. Choose "I don't
know" if he/she does not have an opportunity to play with other cats}
\item{Runs_unfamiliar_dogs}{Runs or hides from unfamiliar dogs.}
\item{waiting_household_member_arrives}{Always comes to the hallway/next
to front door or is already waiting there when a household member
arrives home.}
\item{sudden_licking_chewing_body}{Exhibits sudden frantic licking or
chewing his/her body.}
\item{Comes_called}{Comes when called.}
\item{Escapes_threatened_familiar_cat}{Escapes or freezes when growled at,
hissed at or otherwise threatened by a familiar cat.}
\item{Chews_cables}{Chews cables and electric wires.}
\item{Reacts_different_events_vocalizing}{Reacts to different events and
stimuli by vocalizing (e.g. meowing).}
\item{scratch_bite_familiar_cat_hissing}{Attempts to scratch or bite a
familiar cat when the familiar cat is staring, growling, or hissing
at him/her.}
\item{Growls_hisses_petted_tail}{Growls or hisses when petted at the base
of the tail.}
\item{Chews_plastics}{Chews or licks plastics.}
\item{Growls_hisses_unfamiliar_person_pet}{Growls or hisses when an
unfamiliar person tries to touch or pet him/her.}
\item{Growls_hisses_unfamiliar_children}{Growls or hisses at unfamiliar
children.}
\item{Chase_shadows}{Chases or follows shadows or lights (e.g. laser
pointers, flashlights).}
\item{escape_home}{Attempts to escape the home or yard/garden, if given
the opportunity.}
\item{comfortable_petted_unfamiliar_people}{Is comfortable and relaxed
when being petted by unfamiliar people.}
\item{greets_unfamiliar_adults}{Always greets unfamiliar adults visiting
your home in a friendly manner (e.g. sniffs, rubs, purrs, lies on
the floor).}
\item{Active}{Active, does not spend long time idle.}
\item{Prefers_types_cat_litter}{Prefers specific types of cat litter.}
\item{restlessness_member_household_affection_another_pet}{Shows
restlessness (e.g. vocalizes) when a member of the household shows
affection to another pet.}
\item{Seeks_physical_contact_other_cats}{Seeks physical contact from other
cats in the household, e.g. sleeps next to them or grooms them.
Choose “I don't know" if he/she is the only cat in the household.}
\item{Growls_hisses_approached_familiar_cat_resting}{Growls or hisses when
approached by a familiar cat in a favorite resting place.}
\item{Friendly_unfamiliar_dogs}{Friendly to unfamiliar dogs.}
\item{Growls_hisses_approached_human_possession_of_prey}{Growls or hisses
when approached by human while in possession of prey, stolen or
especially delicious food.}
\item{finds_new_ways_get_attention}{Quickly finds new ways to get
attention.}
\item{learns_associate_things}{Quickly learns to associate separate things
(e.g. opening the fridge means getting food).}
\item{Concentration_not_disturbed}{Concentration is not disturbed by
anything (e.g. other cats) during a training session. Choose "I
don't know" if you do not train your cat.}
\item{Sucks_chews_fabrics}{Sucks on or chews soft fabrics or wool.}
\item{Growls_hisses_unfamiliar_adults}{Growls or hisses at unfamiliar
adults.}
\item{Purrs_sitting_lap}{Purrs when sitting on someone’s lap (has jumped
up voluntarily).}
\item{Chases_bite_moving_legs}{Chases or bites people's moving legs or
feet (in a non-playful way).}
\item{not_recognize_familiar_cat_if_smells_odd}{Does not seem to recognize
(e.g. hisses or escapes) a familiar cat if it smells odd (e.g. the
familiar cat has visited a vet clinic). Choose "I don't know" if
he/she is the only cat in the household.}
\item{Kneads_objects}{Kneads on soft toys, fabrics, and other objects.}
\item{Squirms_picked_to_lap}{Squirms and tries to escape when picked up to
the lap of a sitting/lying person.}
\item{substrate_preference_inappro}{Shows substrate preference when
urinating or defecating in inappropriate places (e.g. laundry,
bedding, carpets, flower pots). Choose "I don't know" if he/she does
not urinate or defecate in inappropriate places.}
\item{interactive_play_people}{Initiates interactive play with people (e.g.
brings toys, strings, or candy wrappers to play with).}
\item{Repeats_meaningless_movements}{Repeats the same meaningless movements
(e.g. pacing back and forth next to a wall).}
\item{comfortable_playing_unfamiliar_people}{Appears comfortable
(confident, relaxed) when playing with unfamiliar people.}
\item{Walks_tail_upright}{Walks tail held upright at home.}
\item{not_approach_people_right_away}{Does not approach people visiting
his/her home right away (he/she is suspicious; however, he/she might
greet people later on).}
\item{Rests_elevated_places}{Rests or sleeps in elevated places (e.g.
shelves, bookcases, tops of wardrobes or cupboards).}
\item{comfortable_unfamiliar_places}{Seems comfortable and confident in
unfamiliar places outside the home (e.g. cat shows).}
\item{responds_commands}{Readily responds to commands taught to him/her
(e.g. sit, down, paw). Choose "I don't know" if you have not trained
him/her.}
\item{Picky}{Picky: only eats foods he/she likes and leaves other foods
untouched.}
\item{Asks_food}{Asks for food when hungry (e.g. meows next to his/her
bowl).}
\item{restlessness_household_member_leave}{Shows restlessness or pacing
when a household member prepares to leave the home.}
\item{exited_other_pets}{Is fascinated by the activities of other small
pets (e.g. rodents, birds, reptiles, fish) in the home.}
\item{Human_oriented}{Human-oriented: Enjoys the company of people more
than the company of other cats.}
}
}

\subsection{Factor Scores}{

\describe{

All values are continuous.

\item{fearfulness}{fearfulness factor score, with higher score indicating
higher fearfulness.}
\item{human_aggression}{aggression toward humans factor score, with higher
score indicating higher aggression toward humans.}
\item{activity_playfulness}{activty/playfulness factor score, with higher
score indicating higher activity/playfulness.}
\item{cat_sociability}{sociability toward cats factor score, with higher
score indicating higher sociability toward cats.}
\item{human_sociability}{sociability toward humans factor score, with
higher score indicating higher sociability toward humans}
\item{litterbox_issues}{litterbox issues factor score, with higher score
indicating more litterbox issues.}
\item{excessive_grooming}{excessive grooming factor score, with higher
score indicating more excessive grooming.}
}
}
}
\source{
\url{https://www.mdpi.com/2076-2615/11/7/1991}

\url{https://figshare.com/articles/dataset/Feline_behavior_and_personality_survey_data/14899077}
}
\usage{
cat_survey
}
\description{
Data set, containing demographic and background information, combined with
questionaire answers, followed by factor scores derived by answered.
}
\references{
Mikkola S, Salonen M, Hakanen E, Sulkama S, Lohi H. Reliability and Validity
of Seven Feline Behavior and Personality Traits. Animals. 2021; 11(7):1991.
https://doi.org/10.3390/ani11071991
}
\keyword{datasets}
